+++ plugins/azure-wakamiti-plugin/src/main/java/es/iti/wakamiti/azure/api/model/PointAssignment.java
@@ -9,6 +9,8 @@ package es.iti.wakamiti.azure.api.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.util.Objects;
+
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class PointAssignment extends BaseModel {
@@ -55,4 +57,8 @@ public class PointAssignment extends BaseModel {
         return this;
     }
 
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(id);
+    }
 }