+++ plugins/azure-wakamiti-plugin/src/main/java/es/iti/wakamiti/azure/api/model/query/criteria/Criteria.java
@@ -203,6 +203,7 @@ public class Criteria implements Expression {
             this.string = string;
         }
 
+        @Override
         public String toString() {
             return string;
         }