+++ wakamiti-api/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ### Changed
-- Abstract HttpClient
+- Abstract HttpClient.
 
 
 ## [2.5.0] - 2024-09-09